Adding Core to Backend Connection

+++ Backend/src/main/java/net/villagerzock/backend/dto/NetworkPointDto.java
@@ -1,4 +1,4 @@
 package net.villagerzock.backend.dto;
 
-public record NetworkPointDto(String key, int inbound, int outbound) {
+public record NetworkPointDto(String key, long inbound, long outbound) {
 }
